requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
diff-match-patch==20241021
dj-database-url==3.0.0
Django==5.2.3
django-import-export==4.3.7
django-jazzmin==3.0.1
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.5.0
oscrypto==1.3.0
pdf2image==1.17.0
pillow==11.2.1
propcache==0.3.2
PyAutoGUI==0.9.54
pycparser==2.22
pydyf==0.11.0
PyGetWindow==0.0.9
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMsgBox==1.0.9
pypdf==5.6.1
pyperclip==1.9.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.6.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
xhtml2pdf==0.2.17
yarl==1.20.1
zopfli==0.2.3.post1
